Formalise `passthrough` behaviour for mocking.
When mocking, a common requirement is for the mock to retain all other
functions in the original module. In order to achieve this behaviour in
meck/Pavlov you must use the `passthrough` option. A test has been
included that formalises this behaviour, and the README has been updated
to document this behaviour.